refactor(scripts): converted the createComponent script to an ES6 module and stopped generating the index.js file
