+++ package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview",
     "test": "vitest",
     "test:ci": "npm run lint && vitest run",
-    "create-component": "node scripts/createComponent.js",
+    "create-component": "node scripts/createComponent.mjs",
     "lint": "eslint ./src/ ./scripts/ --ext .js,.jsx",
     "docker:build": "docker build -t istex/istex-dl:$npm_package_version --build-arg http_proxy --build-arg https_proxy . && docker image prune --filter label=stage=builder -f",
     "docker:run": "docker-compose up -d",
